+++ sdk/storage/azure-storage-common/inc/azure/storage/common/storage_switch_to_secondary_policy.hpp
